finished script and added documentation
